+++ src/viewpoint-measures/SaliencyEVMI.cpp
@@ -32,7 +32,7 @@ void SaliencyEVMI::Compute(const SceneInformationBuilder *pSceneInformationBuild
         {
             Debug::Log("SaliencyEVMI::No neighbours");
         }
-        for( int currentNeighbour = 0; currentNeighbour < neighbours.size(); currentNeighbour++ )
+        for( size_t currentNeighbour = 0; currentNeighbour < neighbours.size(); currentNeighbour++ )
         {
             unsigned int sum_a_z_i = projectedAreasMatrix->GetSumPerPolygon(currentPolygon);
             unsigned int sum_a_z_j = projectedAreasMatrix->GetSumPerPolygon(neighbours.at(currentNeighbour));
